% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asimetria.R
\name{asimetry}
\alias{asimetry}
\title{Asimetria}
\usage{
asimetry(x)
}
\arguments{
\item{x}{Grupo de datos numericos.}
}
\value{
El valor de la asimetria y su error estandar
}
\description{
Una funcion que nos permite sacar el valor de la asimetria y su error estandar
usando funciones que podemos encontrar en los paquetes \code{\link{psych}} y \code{sur}.
}
\examples{
asimetry(c(10,2,35,43))
}
